tcpConnection manualed
