+++ Fedya/FANUC Client/FANUC Client.vcxproj
@@ -146,6 +146,7 @@
     <ClCompile Include="Main.cpp" />
     <ClCompile Include="MyQueue.cpp" />
     <ClCompile Include="MyThread.cpp" />
+    <ClCompile Include="tcpConnection.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="MyQueueDifintion.hpp" />
@@ -160,6 +161,7 @@
     <ClInclude Include="RobotSend.hpp" />
     <ClInclude Include="SocketWorking.hpp" />
     <ClInclude Include="tcpConnection.hpp" />
+    <ClInclude Include="tcpConnectionDifinition.hpp" />
     <ClInclude Include="tcpServer.hpp" />
   </ItemGroup>
   <ItemGroup>
